Move JP migration screen components to generic package
