+++ WordPress/src/main/java/org/wordpress/android/ui/compose/components/Subtitle.kt
@@ -1,26 +1,26 @@
-package org.wordpress.android.ui.main.jetpack.migration.compose.components
+package org.wordpress.android.ui.compose.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import org.wordpress.android.R.dimen
-import org.wordpress.android.ui.compose.unit.FontSize
 
 @Composable
-fun Title(
+fun Subtitle(
     text: String,
     modifier: Modifier = Modifier,
 ) {
     Text(
         text = text,
-        fontSize = FontSize.ExtraExtraExtraLarge.value,
-        fontWeight = FontWeight.Bold,
+        fontSize = 17.sp,
+        style = TextStyle(letterSpacing = (-0.01).sp),
         modifier = modifier
             .padding(horizontal = dimensionResource(dimen.jp_migration_padding_horizontal))
-            .padding(top = 30.dp)
+            .padding(top = 20.dp)
     )
 }
