+++ WordPress/src/main/java/org/wordpress/android/ui/main/jetpack/migration/compose/state/DeleteStep.kt
@@ -23,18 +23,18 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.wordpress.android.R
+import org.wordpress.android.ui.compose.components.ButtonsColumn
+import org.wordpress.android.ui.compose.components.PrimaryButton
+import org.wordpress.android.ui.compose.components.ScreenIcon
+import org.wordpress.android.ui.compose.components.SecondaryButton
+import org.wordpress.android.ui.compose.components.Subtitle
+import org.wordpress.android.ui.compose.components.Title
 import org.wordpress.android.ui.compose.theme.AppTheme
 import org.wordpress.android.ui.compose.utils.htmlToAnnotatedString
 import org.wordpress.android.ui.compose.utils.uiStringText
 import org.wordpress.android.ui.main.jetpack.migration.JetpackMigrationViewModel.ActionButton.DeletePrimaryButton
 import org.wordpress.android.ui.main.jetpack.migration.JetpackMigrationViewModel.ActionButton.DeleteSecondaryButton
 import org.wordpress.android.ui.main.jetpack.migration.JetpackMigrationViewModel.UiState
-import org.wordpress.android.ui.main.jetpack.migration.compose.components.ButtonsColumn
-import org.wordpress.android.ui.main.jetpack.migration.compose.components.PrimaryButton
-import org.wordpress.android.ui.main.jetpack.migration.compose.components.ScreenIcon
-import org.wordpress.android.ui.main.jetpack.migration.compose.components.SecondaryButton
-import org.wordpress.android.ui.main.jetpack.migration.compose.components.Subtitle
-import org.wordpress.android.ui.main.jetpack.migration.compose.components.Title
 
 @Composable
 fun DeleteStep(uiState: UiState.Content.Delete) = with(uiState) {
